requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.5.0
python-dotenv>=1.0.0
aiogram>=3.0.0
numpy>=1.24.0
asyncio>=3.4.3